Add files via upload
Any website requires SSL (Secure Sockets Layer) certificates to protect user data, confirm the site's ownership, prevent hackers from building a fake version of the site, and inspire trust in users. So, in this project we have to configure SSL certificate and verify a website. Need to generate padlock icon which defines that connection is secure on that specific website. Configured CSR Configuration and Generation for the www.verysecureserver.com. Then transfereed CSR to AcmeCA and AcmeCA to www.verysecureserver.com. Then we installed the certificates and made an simple file uploading html page named index.html which we made secured. Then, we configured DNS to access the secured site from other servers and finally to protect our website from unwanted access we configured firewall. That’s how we verified the security of the connection in ubuntu and finally revoke the certificate.